# package-search: improving performance on fetching facet display name