@alexandru-m-g 2015-10-18
Weird that this test failed: `ckanext/datastore/tests/test_create.py` on python 2.7. It didn't fail on python 2.6 and before, the same python 2.7 test, worked: https://travis-ci.org/ckan/ckan/jobs/86087573. Also locally the test passes. 
Will close this pull request and will try again.